@@ -16,43 +23,61 @@
         "test:unit": "vitest",
         "mkcert": "cd certificates && mkcert create-ca && npx mkcert create-cert --domains local-web.scytta.com,local-web.scytta.com:4443"
     },
-    "devDependencies": {
-        "@playwright/test": "^1.45.3",
-        "@sveltejs/adapter-auto": "^3.0.0",
-        "@sveltejs/adapter-cloudflare": "^4.7.0",
-        "@sveltejs/kit": "^2.0.0",
-        "@sveltejs/vite-plugin-svelte": "^3.0.0",
-        "@tailwindcss/nesting": "^0.0.0-insiders.565cd3e",
-        "@tailwindcss/typography": "^0.5.13",
-        "@types/debug": "^4.1.12",
-        "@types/eslint": "^9.6.0",
-        "@types/node": "^22.0.2",
-        "autoprefixer": "^10.4.19",
-        "daisyui": "^4.12.10",
-        "eslint": "^9.0.0",
-        "eslint-config-prettier": "^9.1.0",
-        "eslint-plugin-svelte": "^2.36.0",
-        "globals": "^15.0.0",
-        "husky": "^9.1.4",
-        "mkcert": "^3.2.0",
-        "postcss": "^8.4.40",
-        "postcss-import": "^16.1.0",
-        "prettier": "^3.1.1",
-        "prettier-plugin-svelte": "^3.1.2",
-        "prettier-plugin-tailwindcss": "^0.6.4",
-        "svelte": "^5.0.0-next.201",
-        "svelte-check": "^3.6.0",
-        "tailwindcss": "^3.4.7",
-        "typescript": "^5.0.0",
-        "typescript-eslint": "^8.0.0-alpha.20",
-        "vite": "^5.0.3",
-        "vitest": "^2.0.0",
-        "wrangler": "^3.70.0"
-    },
-    "type": "module",
     "dependencies": {
-        "debug": "^4.3.6",
-        "runed": "^0.15.0",
-        "sveltekit-i18n": "^2.4.2"
+        "@floating-ui/dom": "^1.6.12",
+        "debug": "^4.3.7",
+        "sveltekit-i18n": "^2.4.2",
+        "tailwind-merge": "^2.5.5",
+        "turnstile-types": "^1.2.3",
+        "ua-parser-js": "^2.0.0",
+        "uuid": "^11.0.3",
+        "zod": "^3.23.8"
+    },
+    "devDependencies": {
+        "@playwright/test": "^1.49.0",
+        "@skeletonlabs/tw-plugin": "^0.4.0",
+        "@sveltejs/adapter-auto": "^3.3.1",
+        "@sveltejs/adapter-cloudflare": "^4.8.0",
+        "@sveltejs/kit": "^2.9.0",
+        "@sveltejs/vite-plugin-svelte": "^4.0.1",
+        "@tailwindcss/forms": "^0.5.9",
+        "@tailwindcss/nesting": "^0.0.0-insiders.565cd3e",
+        "@trivago/prettier-plugin-sort-imports": "^4.3.0",
+        "@types/debug": "^4.1.12",
+        "@types/eslint": "^9.6.1",
+        "@types/node": "^22.10.1",
+        "@types/ua-parser-js": "^0.7.39",
+        "@types/uuid": "^10.0.0",
+        "autoprefixer": "^10.4.19",
+        "eslint": "^9.16.0",
+        "eslint-config-prettier": "^9.1.0",
+        "eslint-plugin-svelte": "^2.46.1",
+        "globals": "^15.13.0",
+        "husky": "^9.1.7",
+        "mkcert": "^3.2.0",
+        "msw": "^2.6.6",
+        "peggy": "^4.2.0",
+        "postcss": "^8.4.49",
+        "postcss-import": "^16.1.0",
+        "prettier": "^3.4.1",
+        "prettier-plugin-svelte": "^3.3.2",
+        "prettier-plugin-tailwindcss": "^0.6.9",
+        "sharp": "^0.33.5",
+        "svelte": "^5.4.0",
+        "svelte-check": "^4.1.0",
+        "svgo": "^3.3.2",
+        "tailwindcss": "^3.4.15",
+        "tsx": "^4.19.2",
+        "typescript": "^5.7.2",
+        "typescript-eslint": "^8.17.0",
+        "vite": "^5.4.11",
+        "vite-plugin-static-copy": "^2.2.0",
+        "vitest": "^2.1.8",
+        "wrangler": "^3.91.0"
+    },
+    "msw": {
+        "workerDirectory": [
+            "static-generated"
+        ]
     }
 }
