@@ -1,8 +1,8 @@
 import js from '@eslint/js';
-import ts from 'typescript-eslint';
-import svelte from 'eslint-plugin-svelte';
 import prettier from 'eslint-config-prettier';
+import svelte from 'eslint-plugin-svelte';
 import globals from 'globals';
+import ts from 'typescript-eslint';
 
 /** @type {import('eslint').Linter.Config[]} */
 export default [
@@ -14,6 +14,7 @@ export default [
     {
         languageOptions: {
             globals: {
+                ENABLE_MOCK: 'readonly',
                 ...globals.browser,
                 ...globals.node
             }
@@ -25,9 +26,12 @@ export default [
             parserOptions: {
                 parser: ts.parser
             }
+        },
+        rules: {
+            '@typescript-eslint/no-unused-vars': ['error', { argsIgnorePattern: '^_|_' }]
         }
     },
     {
-        ignores: ['build/', '.svelte-kit/', 'dist/']
+        ignores: ['build/', '.svelte-kit/', 'dist/', 'static-generated/']
     }
 ];
