@@ -1,4 +1,4 @@
-import { describe, it, expect } from 'vitest';
+import { describe, expect, it } from 'vitest';
 
 describe('sum test', () => {
     it('adds 1 + 2 to equal 3', () => {
