@@ -1,4 +1,6 @@
-<div class="error">
-    <p>Something went wrong 2</p>
-    <a href="/design">Back</a>
-</div>
+<script>
+    import { page } from '$app/stores';
+    import ErrorCard from '$atoms/ErrorCard.svelte';
+</script>
+
+<ErrorCard error={{ errorKind: 'other', message: $page.error?.message ?? '' }} />
