@@ -1,2 +1,7 @@
-Nothing to see here
-<a href="/design/foo">Foo</a>
+<script>
+    import Typography from '$atoms/Typography.svelte';
+</script>
+
+<div class="flex h-full items-center justify-around">
+    <Typography variant="h1" element="p">Nothing to see here</Typography>
+</div>
