@@ -1,7 +1,7 @@
-import { identityApi } from '$lib/api/identity_api';
+import { identityApi } from '$lib/api/identity-api';
 
 export const load = async ({ fetch }) => {
     const providers = await identityApi.getExternalLoginProviders(fetch);
-    console.log('after load:', providers);
+
     return { providers };
 };
